requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-telegram-bot==20.8
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
pydantic==2.8.2
httpx==0.26.0
tenacity==8.5.0
redis==5.0.7
python-dotenv==1.0.1
tronpy==0.4.0
cryptography==42.0.8
openai==1.37.1
tqdm==4.66.4
bitcoinlib==0.6.15
bitcoinlib==0.6.15
